Move diff to subpackage
Reduces clutter
