+++ src/git_autograder/diff/diff_helper.py
@@ -1,20 +1,10 @@
-from dataclasses import dataclass
-from typing import Iterator, Optional
+from typing import Iterator
 
 from difflib_parser.difflib_parser import DiffParser
-from git import Commit, DiffIndex
-from git.diff import Diff, Lit_change_type
+from git import Commit, Diff, DiffIndex
+from git.diff import Lit_change_type
 
-
-@dataclass
-class GitAutograderDiff:
-    diff: Diff
-    change_type: Lit_change_type
-    original_file_path: Optional[str]
-    edited_file_path: Optional[str]
-    original_file: Optional[str]
-    edited_file: Optional[str]
-    diff_parser: Optional[DiffParser]
+from git_autograder.diff.diff import GitAutograderDiff
 
 
 class GitAutograderDiffHelper:
